+++ lib/actions/grants/client_credentials.js
@@ -4,15 +4,15 @@ const instance = require('../../helpers/weak_cache');
 module.exports.handler = function getClientCredentialsHandler(provider) {
   return async function clientCredentialsResponse(ctx, next) {
     const { ClientCredentials } = provider;
-    const scope = intersection(
-      String(ctx.oidc.params.scope).split(' '),
-      instance(provider).configuration('scopes'),
-    );
+    const { audiences, scopes } = instance(provider).configuration();
+    const scope = intersection(String(ctx.oidc.params.scope).split(' '), scopes).join(' ');
 
     const token = new ClientCredentials({
       clientId: ctx.oidc.client.clientId,
-      scope: scope.length ? scope.join(' ') : undefined,
+      scope: scope || undefined,
     });
+
+    token.setAudiences(await audiences(ctx, undefined, token, 'client_credentials', scope));
 
     const value = await token.save();
     ctx.oidc.entity('ClientCredentials', token);
