+++ lib/models/access_token.js
@@ -1,16 +1,5 @@
-const ensureConform = require('../helpers/ensure_conform');
+const setAudiences = require('./mixins/set_audiences');
 
 module.exports = function getAccessToken({ BaseToken }) {
-  return class AccessToken extends BaseToken {
-    setAudiences(audiences) {
-      const { clientId } = this;
-      if (audiences) {
-        const value = ensureConform(audiences, clientId);
-
-        if (value.length > 1) {
-          this.aud = value;
-        }
-      }
-    }
-  };
+  return class AccessToken extends setAudiences(BaseToken) {};
 };
