+++ lib/models/client_credentials.js
@@ -1,3 +1,5 @@
+const setAudiences = require('./mixins/set_audiences');
+
 module.exports = function getClientCredentials({ BaseToken }) {
-  return class ClientCredentials extends BaseToken {};
+  return class ClientCredentials extends setAudiences(BaseToken) {};
 };
