+++ lib/models/mixins/set_audiences.js
@@ -1,0 +1,14 @@
+const ensureConform = require('../../helpers/ensure_conform');
+
+module.exports = superclass => class extends superclass {
+  setAudiences(audiences) {
+    const { clientId } = this;
+    if (audiences) {
+      const value = ensureConform(audiences, clientId);
+
+      if (value.length > 1) {
+        this.aud = value;
+      }
+    }
+  }
+};
